Improve plotting
Ability to plot multiple regions together, or even data from multiple
tracks. Optionally saved to PDF. Backwards incompatible change is that
the command line plotting function has been renamed from *visualize* to
*plot*.

The main plotting functionality is in its own module, so it can also be
used as a library.